Add files via upload
To be used efficiently, all computer software needs certain hardware components or the
other software resources to be present on a computer. These pre requisites are
known as(computer) system requirements and are often used as a guideline as
opposed to an absolute rule. Most software defines two sets of system
requirements: minimum and recommended. With increasing demand for higher
processing power and resources in newer versions of software, system requirements
tend to increase over time. Industry analysts suggest that this trend plays a bigger
part in driving upgrades to existing computer systems than technological
advancements.